src/groq-llm-api/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
langchain==0.3.27
langchain-core==0.3.74
langchain-groq==0.3.7
python-dotenv==1.1.1